Epicrop

Documentation for Epicrop.

Epicrop.bacterialblight

Epicrop.brownspot

Epicrop.hlipmodel

Epicrop.leafblast

Epicrop.sheathblight

Epicrop.tungro

Epicrop.bacterialblight — Method

bacterialblight(wth::DataFrames.AbstractDataFrame, emergence::Dates.Date)

Run a healthy-latent-infectious-postinfectious (HLIP) model using weather data and optimal curve values for rice bacterial blight Xanthomonas oryzae pv. oryzae.

Arguments

wth: DataFrames:DataFrame a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD: Date as Year Month Day, YYYY-MM-DD, (ISO8601)

DOY`: Consecutive day of year, commonly called "Julian date"

TEMP: Mean daily temperature (°C)`

RHUM: Mean daily relative humidity (%)

RAIN: Mean daily rainfall (mm)

emergence: expected date of plant emergence as a Date object. From Table 1 Savary et al. 2012.

Examples

julia> weather=
julia> emergence=Dates.Date.("2010-07-01", Dates.DateFormat("yyyy-mm-dd"))
julia> bacterialblight(weather, emergence)

Returns

A Table with predictions for bacterial blight severity. Latitude and longitude are included for mapping purposes if they are present in the input weather data.

Epicrop.brownspot — Method

brownspot(wth::DataFrames.AbstractDataFrame, emergence::Dates:Date)

Run a healthy-latent-infectious-postinfectious (HLIP) model using weather data and optimal curve values for rice brown spot caused by Cochliobolus miyabeanus.

Arguments

wth: DataFrames:DataFrame a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD: Date as Year Month Day, YYYY-MM-DD, (ISO8601)

DOY`: Consecutive day of year, commonly called "Julian date"

TEMP: Mean daily temperature (°C)`

RHUM: Mean daily relative humidity (%)

RAIN: Mean daily rainfall (mm)

emergence: expected date of plant emergence as a Date object. From Table 1 Savary et al. 2012.

Examples

julia> weather=
julia> emergence=Dates.Date.("2010-07-01", Dates.DateFormat("yyyy-mm-dd"))
julia> brownspot(weather, emergence)

Returns

A Table with predictions for brown spot severity. Latitude and longitude are included for mapping purposes if they are present in the input weather data.

Epicrop.hlipmodel — Method

hlipmodel(
    wth::DataFrames.AbstractDataFrame,
    emergence::Dates.Date,
    onset::Int,
    duration::Int,
    rhlim::Int,
    rainlim::Int,
    H0::Int,
    I0::Int,
    RcA::Matrix{AbstractFloat},
    RcT::Matrix{AbstractFloat},
    RcOpt::AbstractFloat,
    p::Int,
    i::Int,
    Sx::Int,
    a::AbstractFloat,
    RRS::AbstractFloat,
    RRG::AbstractFloat,
)

Run a healthy-latent-infectious-postinfectious (HLIP) model using weather data and optimal curve values for respective crop diseases.

Keywords

wth: DataFrames:DataFrame a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD: Date as Year Month Day, YYYY-MM-DD, (ISO8601)

DOY`: Consecutive day of year, commonly called "Julian date"

TEMP: Mean daily temperature (°C)`

RHUM: Mean daily relative humidity (%)

RAIN: Mean daily rainfall (mm)

emergence: Dates.Date expected date of plant emergence entered as a Dates.Date object. From Table 1 Savary et al. 2012.

onset: Int, expected number of days until the onset of disease after emergence date. From Table 1 Savary et al. 2012.

duration: Int, simulation duration (growing season length). From Table 1 Savary et al. 2012.

rhlim: Int, threshold to decide whether leaves are wet or not (usually 90%). From Table 1 Savary et al. 2012.

rainlim: Int, threshold to decide whether leaves are wet or not. From Table 1 Savary et al. 2012.

H0: Int, initial number of plant's healthy sites. From Table 1 Savary et al. 2012.

I0: Int, initial number of infective sites. From Table 1 Savary et al. 2012.

RcA: Matrix{AbstractFloat}, crop age modifier for Rc (the basic infection rate corrected for removals). From Table 1 Savary et al. 2012.

RcT: Matrix{AbstractFloat}, temperature modifier for Rc (the basic infection rate corrected for removals). From Table 1 Savary et al. 2012.

RcOpt: AbstractFloat, potential basic infection rate corrected for removals. From Table 1 Savary et al. 2012.

p: Int, duration of latent period. From Table 1 Savary et al. 2012.

Sx: Int, maximum number of sites. From Table 1 Savary et al. 2012.

a: AbstractFloat, aggregation coefficient. From Table 1 Savary et al. 2012.

RRS: AbstractFloat, relative rate of physiological senescence. From Table 1 Savary et al. 2012.

RRG: AbstractFloat, relative rate of growth. From Table 1 Savary et al. 2012.

Returns

A DataFrame with the model's output. Latitude and longitude are included for mapping purposes if they are present in the input weather data.

Epicrop.leafblast — Method

leafblast(wth::DataFrames.AbstractDataFrame, emergence::Dates.Date)

Run a healthy-latent-infectious-postinfectious (HLIP) model using weather data and optimal curve values for rice leaf blast caused by Magnaporthe oryzae.

Arguments

wth: DataFrames:DataFrame a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD: Date as Year Month Day, YYYY-MM-DD, (ISO8601)

DOY`: Consecutive day of year, commonly called "Julian date"

TEMP: Mean daily temperature (°C)`

RHUM: Mean daily relative humidity (%)

RAIN: Mean daily rainfall (mm)

emergence: expected date of plant emergence as a Date object. From Table 1 Savary et al. 2012.

Examples

julia> weather=
julia> emergence=Dates.Date.("2010-07-01", Dates.DateFormat("yyyy-mm-dd"))
julia> leafblast(weather, emergence)

Returns

A Table with predictions for leaf blast severity. Latitude and longitude are included for mapping purposes if they are present in the input weather data.

Epicrop.sheathblight — Method

sheathblight(wth::DataFrames.AbstractDataFrame, emergence::Dates.Date)

Run a healthy-latent-infectious-postinfectious (HLIP) model using weather data and optimal curve values for rice sheath blight caused by Rhizoctonia solani AG1-1A Kühn.

Arguments

wth: DataFrames:DataFrame a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD: Date as Year Month Day, YYYY-MM-DD, (ISO8601)

DOY`: Consecutive day of year, commonly called "Julian date"

TEMP: Mean daily temperature (°C)`

RHUM: Mean daily relative humidity (%)

RAIN: Mean daily rainfall (mm)

emergence: expected date of plant emergence as a Date object. From Table 1 Savary et al. 2012.

Examples

julia> weather=
julia> emergence=Dates.Date.("2010-07-01", Dates.DateFormat("yyyy-mm-dd"))
julia> sheathblight(weather, emergence)

Returns

A Table with predictions for sheath blight severity. Latitude and longitude are included for mapping purposes if they are present in the input weather data.

Epicrop.tungro — Method

tungro(wth::DataFrames.AbstractDataFrame, emergence::Dates.Date)

Run a healthy-latent-infectious-postinfectious (HLIP) model using weather data and optimal curve values for rice tungro disease caused by Rice Tungro Spherical and Rice Tungro Bacilliform viruses.

Arguments

wth: DataFrames:DataFrame a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD: Date as Year Month Day, YYYY-MM-DD, (ISO8601)

DOY`: Consecutive day of year, commonly called "Julian date"

TEMP: Mean daily temperature (°C)`

RHUM: Mean daily relative humidity (%)

RAIN: Mean daily rainfall (mm)

emergence: expected date of plant emergence as a Date object. From Table 1 Savary et al. 2012.

Examples

julia> weather=
julia> emergence=Dates.Date.("2010-07-01", Dates.DateFormat("yyyy-mm-dd"))
julia> tungro(weather, emergence)

Returns

A Table with predictions for tungro incidence. Latitude and longitude are included for mapping purposes if they are present in the input weather data.