Epicrop

Documentation for Epicrop.

Epicrop.bacterialblight

Epicrop.brownspot

Epicrop.hlipmodel

Epicrop.leafblast

Epicrop.sheathblight

Epicrop.tungro

Helper Functions

Quickly and easily model bacterial blight, brown spot, leaf blast, sheath blight and tungro. These functions provide the basic values necessary as published in Savary et al. (2012) to model these diseases with just daily weather data that spans the growing season of interest and growing season start date as values.

Bacterial Blight

A dynamic mechanistic simulation of bacterial blight disease of rice, causal agent Xanthomonas oryzae pv. oryzae. The model represents site size as 1 rice plant's leaf. Default values for this disease model are derived from Table 2 (Savary et al. 2012).

Epicrop.bacterialblight — Function

bacterialblight(wth::DataFrames.AbstractDataFrame, emergence::Dates.Date)

Run a healthy-latent-infectious-postinfectious (HLIP) model using weather data and optimal curve values for rice bacterial blight Xanthomonas oryzae pv. oryzae.

Input

wth: DataFrames.DataFrame a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD: Date as Year Month Day, YYYY-MM-DD, (ISO8601) format either as an AbstractString or Dates.Date

DOY: Consecutive day of year, commonly called "Julian date" as an Int64

TEMP: Mean daily temperature (°C) as aa Float64

RHUM: Mean daily relative humidity (%) as a Float64

RAIN: Mean daily rainfall (mm) as a Float64

emergence: expected date of plant emergence as an AbstractString. From Table 1 Savary et al. 2012.

duration: duration of the growing season in days as an Int64 value, defaults to 120 as in Savary et al. 2012.

Output

A DataFrames.DataFrame with predictions for bacterial blight severity. Latitude and longitude are included for mapping purposes if they are present in the input weather data. See hlipmodel for a full description of the return values.

Notes

This is one of a family of helper functions. Also see brownspot, leafblast, sheathblight and tungro.

Examples

julia> using Epicrop

julia> using DelimitedFiles

julia> using DataFrames

julia> data, header = readdlm(joinpath(
                                dirname(pathof(Epicrop)),
                                    "..", "docs", "src", "assets",
                                        "POWER_data_LB_PHI_2000_ws.csv"),
                                ',', header=true)

julia> w = DataFrame(data, vec(header))

julia> emergence = "2000-07-01"

julia> bacterialblight(w, emergence)

Brown Spot

A dynamic mechanistic simulation of rice brown spot, causal agent Cochliobolus miyabeanus. The model represents site size as 10 mm² of a rice plant's leaf. Default values for this disease model are derived from Table 2 (Savary et al. 2012).

Note The optimum temperature for brown spot as presented in Table 2 of Savary et al. 2012 has a typo. The optimal value should be 25 °C, not 20 °C as shown. The correct value, 25 °C, is used in this implementation.

Epicrop.brownspot — Function

brownspot(wth::DataFrames.AbstractDataFrame, emergence::Dates:Date)

Run a healthy-latent-infectious-postinfectious (HLIP) model using weather data and optimal curve values for rice brown spot caused by Cochliobolus miyabeanus.

Input

wth: DataFrames.DataFrame a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD: Date as Year Month Day, YYYY-MM-DD, (ISO8601) format either as an AbstractString or Dates.Date

DOY: Consecutive day of year, commonly called "Julian date" as an Int64

TEMP: Mean daily temperature (°C) as aa Float64

RHUM: Mean daily relative humidity (%) as a Float64

RAIN: Mean daily rainfall (mm) as a Float64

emergence: expected date of plant emergence as an AbstractString. From Table 1 Savary et al. 2012.

duration: duration of the growing season in days as an Int64 value, defaults to 120 as in Savary et al. 2012.

Output

A DataFrames.DataFrame with predictions for brown spot severity. Latitude and longitude are included for mapping purposes if they are present in the input weather data. See hlipmodel for a full description of the return values.

This function is one of a family of helper functions. Also see bacterialblight, leafblast, sheathblight and tungro.

Examples

julia> using Epicrop

julia> using DelimitedFiles

julia> using DataFrames

julia> data, header = readdlm(joinpath(
                                dirname(pathof(Epicrop)),
                                    "..", "docs", "src", "assets",
                                        "POWER_data_LB_PHI_2000_ws.csv"),
                                ',', header=true)

julia> w = DataFrame(data, vec(header))

julia> emergence = "2000-07-01"

julia> brownspot(w, emergence)

Leaf Blast

A dynamic mechanistic simulation of leaf blast disease of rice, causal agent Magnaporthe oryzae. The model represents site size as 45 mm² of a rice plant's leaf.

Note The optimum temperature for leaf blast as presented in Table 2 of Savary et al. 2012 has a typo. The optimal value should be 20 °C, not 25 °C as shown. The correct value, 20 °C, is used in this implementation.

Epicrop.leafblast — Function

leafblast(wth::DataFrames.AbstractDataFrame, emergence::Dates.Date)

Run a healthy-latent-infectious-postinfectious (HLIP) model using weather data and optimal curve values for rice leaf blast caused by Magnaporthe oryzae.

Input

wth: DataFrames.DataFrame a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD: Date as Year Month Day, YYYY-MM-DD, (ISO8601) format either as an AbstractString or Dates.Date

DOY: Consecutive day of year, commonly called "Julian date" as an Int64

TEMP: Mean daily temperature (°C) as aa Float64

RHUM: Mean daily relative humidity (%) as a Float64

RAIN: Mean daily rainfall (mm) as a Float64

emergence: expected date of plant emergence as an AbstractString. From Table 1 Savary et al. 2012.

duration: duration of the growing season in days as an Int64 value, defaults to 120 as in Savary et al. 2012.

Output

A DataFrames.DataFrame with predictions for leaf blast severity. Latitude and longitude are included for mapping purposes if they are present in the input weather data. See hlipmodel for a full description of the return values.

Notes

This is one of a family of helper functions. Also see bacterialblight, brownspot, sheathblight and tungro.

Examples

julia> using Epicrop

julia> using DelimitedFiles

julia> using DataFrames

julia> data, header = readdlm(joinpath(
                                dirname(pathof(Epicrop)),
                                    "..", "docs", "src", "assets",
                                        "POWER_data_LB_PHI_2000_ws.csv"),
                                ',', header=true)

julia> w = DataFrame(data, vec(header))

julia> emergence = "2000-07-01"

julia> leafblast(w, emergence)

Sheath Blight

A dynamic mechanistic simulation of sheath blight disease of rice, causal agent Rhizoctonia solani AG1-1A Kühn. The model represents site size as 1 rice plant's tiller.

Epicrop.sheathblight — Function

sheathblight(wth::DataFrames.AbstractDataFrame, emergence::Dates.Date)

Run a healthy-latent-infectious-postinfectious (HLIP) model using weather data and optimal curve values for rice sheath blight caused by Rhizoctonia solani AG1-1A Kühn.

Input

wth: DataFrames.DataFrame a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD: Date as Year Month Day, YYYY-MM-DD, (ISO8601) format either as an AbstractString or Dates.Date

DOY: Consecutive day of year, commonly called "Julian date" as an Int64

TEMP: Mean daily temperature (°C) as aa Float64

RHUM: Mean daily relative humidity (%) as a Float64

RAIN: Mean daily rainfall (mm) as a Float64

emergence: expected date of plant emergence as an AbstractString. From Table 1 Savary et al. 2012.

duration: duration of the growing season in days as an Int64 value, defaults to 120 as in Savary et al. 2012.

Output

A DataFrames.DataFrame with predictions for sheath blight severity. Latitude and longitude are included for mapping purposes if they are present in the input weather data. See hlipmodel for a full description of the return values.

Notes

This is one of a family of helper functions. Also see bacterialblight, brownspot, leafblast and tungro.

Examples

julia> using Epicrop

julia> using DelimitedFiles

julia> using DataFrames

julia> data, header = readdlm(joinpath(
                                dirname(pathof(Epicrop)),
                                    "..", "docs", "src", "assets",
                                        "POWER_data_LB_PHI_2000_ws.csv"),
                                ',', header=true)

julia> w = DataFrame(data, vec(header))

julia> emergence = "2000-07-01"

julia> sheathblight(w, emergence)

Tungro

A dynamic mechanistic simulation of tungro disease of rice, causal agents Rice Tungro Spherical Virus and Rice Tungro Bacilliform Virus. The model represents site size as 1 rice plant. Default values for this disease model are derived from Table 2 (Savary et al. 2012).

Epicrop.tungro — Function

tungro(wth::DataFrames.AbstractDataFrame, emergence::Dates.Date)

Run a healthy-latent-infectious-postinfectious (HLIP) model using weather data and optimal curve values for rice tungro disease caused by Rice Tungro Spherical and Rice Tungro Bacilliform viruses.

Input

wth: DataFrames.DataFrame a data frame of weather on a daily time-step containing data with the following field names.

YYYYMMDD: Date as Year Month Day, YYYY-MM-DD, (ISO8601) format either as an AbstractString or Dates.Date

DOY: Consecutive day of year, commonly called "Julian date" as an Int64

TEMP: Mean daily temperature (°C) as aa Float64

RHUM: Mean daily relative humidity (%) as a Float64

RAIN: Mean daily rainfall (mm) as a Float64

emergence: expected date of plant emergence as an AbstractString. From Table 1 Savary et al. 2012.

duration: duration of the growing season in days as an Int64 value, defaults to 120 as in Savary et al. 2012.

Output

A DataFrames.DataFrame with predictions for tungro incidence. Latitude and longitude are included for mapping purposes if they are present in the input weather data. See hlipmodel for a full description of the return values.

Notes

This is one of a family of helper functions. Also see bacterialblight, brownspot, leafblast and sheathblight.

Examples

julia> using Epicrop

julia> using DelimitedFiles

julia> using DataFrames

julia> data, header = readdlm(joinpath(
                                dirname(pathof(Epicrop)),
                                    "..", "docs", "src", "assets",
                                        "POWER_data_LB_PHI_2000_ws.csv"),
                                ',', header=true)

julia> w = DataFrame(data, vec(header))

julia> emergence = "2000-07-01"

julia> tungro(w, emergence)

# output

Advanced Use

Epicrop.jl provides a basic function, hlipmodel, that can be used to predict unmanaged plant disease epidemics given the proper input. Predefined values for the EPIRICE model can be found in Savary et al. (2012) for the following diseases of rice: bacterial blight, brown spot, leaf blast, sheath blight and tungro and are included as helper functions, bacterialblight, brownspot, leafblast, sheathblight, and tungro, that simplify using the model's interface. Given other parameters, the model framework is capable of modelling other diseases using the methods as described by Savary et al. (2012).

Epicrop.hlipmodel — Function

hlipmodel(;
wth::DataFrames.AbstractDataFrame,
rhlim::Int64,
rainlim::Int64,
H0::Int64,
I0::Int64,
RcA::Matrix{Float64},
RcT::Matrix{Float64},
RcOpt::Float64,
p::Int64,
i::Int64,
Sx::Int64,
a::Float64,
RRS::Float64,
RRG::Float64,
)

Run a healthy-latent-infectious-postinfectious (HLIP) model using weather data and optimal curve values for respective crop diseases.

Input

wth: a Matrix{Any} of weather on a daily time-step containing data with the following values.

YYYYMMDD: Date as in ISO8601 format, i.e., YYYY-MM-DD

DOY: Consecutive day of year, commonly called "Julian date"

TEMP: Mean daily temperature (°C)

RHUM: Mean daily relative humidity (%)

RAIN: Mean daily rainfall (mm)

emergence: Dates.Date expected date of plant emergence entered as a Dates.Date object. From Table 1 Savary et al. 2012.

onset: Integer, expected number of days until the onset of disease after emergence date. From Table 1 Savary et al. 2012.

duration: Integer, simulation duration (growing season length). From Table 1 Savary et al. 2012.

rhlim: Integer, threshold to decide whether leaves are wet or not (usually 90%). From Table 1 Savary et al. 2012.

rainlim: Integer, threshold to decide whether leaves are wet or not. From Table 1 Savary et al. 2012.

H0: Integer, initial number of plant's healthy sites. From Table 1 Savary et al. 2012.

I0: Integer, initial number of infective sites. From Table 1 Savary et al. 2012.

RcA: Matrix{Float64}, crop age modifier for Rc (the basic infection rate corrected for removals). From Table 1 Savary et al. 2012.

RcT: Matrix{Float64}, temperature modifier for Rc (the basic infection rate corrected for removals). From Table 1 Savary et al. 2012.

RcOpt: Float64, potential basic infection rate corrected for removals. From Table 1 Savary et al. 2012.

p: Integer, duration of latent period. From Table 1 Savary et al. 2012.

Sx: Integer, maximum number of sites. From Table 1 Savary et al. 2012.

a: Float64, aggregation coefficient. From Table 1 Savary et al. 2012.

RRS: Float64, relative rate of physiological senescence. From Table 1 Savary et al. 2012.

RRG: Float64, relative rate of growth. From Table 1 Savary et al. 2012.

Output

A Matrix{Any} with the model's output with the following columns and values.

simday: Zero indexed day of simulation that was run.

dates: Date of simulation.

sites: Total number of sites present on day "x".

latent: Number of latent sites present on day "x".

infectious: Number of infectious sites present on day "x".

removed: Number of removed sites present on day "x".

senesced: Number of senesced sites present on day "x".

rateinf: Rate of infection.

rtransfer: Rate of transfer from latent to infectious sites.

rgrowth: Rate of growth of healthy sites.

rsenesced: Rate of senescence of healthy sites.

diseased: Number of diseased (latent + infectious + removed) sites.

intensity: Number of diseased sites as a proportion of total sites.

audpc: Area under the disease progress curve for the whole of simulated season.

lat: Latitude value as provided by wth object.

lon: Longitude value as provided by wth object.

Examples

Using data downloaded and saved from NASA POWER for Los Baños, Calabarzon, PH.

julia> using Epicrop

julia> using CSV

julia> using DelimitedFiles

julia> using DataFrames

julia> data, header = readdlm(joinpath(
                                dirname(pathof(Epicrop)),
                                    "..", "docs", "src", "assets",
                                        "POWER_data_LB_PHI_2000_ws.csv"),
                                ',', header=true)

julia> w = DataFrame(data, vec(header))

julia> emergence = "2000-07-01"

julia> RcA = [0 0.35;
            20 0.35;
            40 0.35;
            60 0.47;
            80 0.59;
            100 0.71;
            120 1.0]

julia> RcT = [15 0.0;
            20 0.06;
            25 1.0;
            30 0.85;
            35 0.16;
            40 0.0]

julia> bs = hlipmodel(;
        wth=w,
        emergence=emergence,
        onset=20,
        duration=120,
        rhlim=90,
        rainlim=5,
        H0=600,
        I0=1,
        RcA=RcA,
        RcT=RcT,
        RcOpt=0.61,
        p=6,
        i=19,
        Sx=100000,
        a=1.0,
        RRS=0.01,
        RRG=0.1
)

# output

References

Savary, S., Nelson, A., Willocquet, L., Pangga, I., and Aunario, J. Modeling and mapping potential epidemics of rice diseases globally. Crop Protection, Volume 34, 2012, Pages 6-17, ISSN 0261-2194 DOI: 10.1016/j.cropro.2011.11.009.